spapr: nested: Add support for P9 compat L2 guest
Currently, the h_guest_get_capabilities H_CALL evaluates the L0
capabilities based on SPR PVR which results the QEMU L0 to either
support the P10 capabilities or the P9 capabilities. Because of this,
it's not possible to boot a P9 compat L2 guest on a P10 L1.

This patch changes the evaluation of the L0 capabilities based on the
available max compat supported by L0 and send both P9 and P10 capabilities
when available; hence adding the support for a P9 compat guest.

This patch further changes the way capabilities are handled in
h_guest_set_capabilities; again based on the available max compat.

Signed-off-by: Amit Machhiwal <amachhiw@linux.vnet.ibm.com>